Add files via upload
In this project, the aim is to build a predictive model for income classification. The goal of this model is to predict whether an individual's income is above or below a certain threshold, based on various demographic and financial features. The model will be trained on a historical dataset and evaluated based on its accuracy and ability to generalize to unseen data. The outcome of this project has the potential to support decision making for a range of applications, such as financial planning, insurance pricing, and credit scoring."